symlinks: fix output for symlinks in various commands
Summary:
Some hg commands (`show`, `cat`, `diff`) were showing backward slashes instead of forward slashes on Windows.

One particular quirk of this was that `hg diff` was not showing the changes for the file if a symlink was replaced for a file with the exact same contents, but after this diff it reverts back to the old behavior.

Reviewed By: quark-zju

Differential Revision: D47979005

fbshipit-source-id: 444974e233a7fcee7fe600d5cb7c7c7ad6d61789