tests: fix bug with nested ifs on debugruntest
Summary:
There was a bug with nested `#if` conditions in debugruntests where code would be tested if the nested condition passed but its parent condition didn't.

For instance, if one had something like

```
#if somethingfalse
...
#if somethingtrue
...
#endif
...
#endif
```

the code under `#if somethingtrue` would be evaluated, which this diff fixes.

This should also allow us to run some previously failing tests such as `test-symlinks.t` with Buck on Watchman on Windows

Reviewed By: zzl0

Differential Revision: D47972193

fbshipit-source-id: 0d18202d9fd9e64b964c56200f35bcb53a9c80a9