hooks: add a dedicated hook for eden redirection
Summary:
The `edenfsctl redirect fixup` hook adds visible overhead especially on Windows
(0.3 to 0.8s).

Let's add a way to run it in background. To avoid breaking automation that relies
on `hg update` to prepare `buck-out`, we still run the redirection fixup in
foreground if we detect obvious issues of the repo redirections.

This does not immediately affect the existing hook behavior.
To use this hook, set:

  [hooks]
  edenfs-update.eden-redirect = python:edenscm.hooks.edenfs_redirect_fixup

Reviewed By: xavierd

Differential Revision: D47808862

fbshipit-source-id: c53566ffa0916a010b06505b44fa8164b870ce91