treemanifest: improve eagerepo compatibility
Summary:
Implement more "history store" and "data store" methods on the EagerRepoStore.
This makes it work better with the treemanifest's "datastore + historystore"
abstractions.

This is needed to be able to reuse logic (ex. unionstore) that deals with the
"bundlerepo" cases.

Note 1: for history we only provide one level history to provide the p1, p2
which is needed when exchange (ex. push) tree contents. We don't follow history
or maintain meaningful linkrevs.

Note 2: It seems the old logic will create "scmstore" for a "revlog" repo,
mixing up shallow and non-shallow (revlog) contents... This change starts to
avoid that.

Reviewed By: zzl0

Differential Revision: D47917532

fbshipit-source-id: baae9d693b2bffdccc40db65595dbb4d21b5b6c2