update test snapshot
Summary:
The `basic-20210712.tar.xz` contains an hg repo that uses the revlog format,
which is being removed. Migrate the repo to also include revlog2 format
(D47878774).

Newly added files are:

  data/repo/.hg/store/hgcommits/
  data/repo/.hg/store/hgcommits/v1/
  data/repo/.hg/store/hgcommits/v1/index2-id
  data/repo/.hg/store/hgcommits/v1/log
  data/repo/.hg/store/hgcommits/v1/meta
  data/repo/.hg/store/hgcommits/v1/rlock
  data/repo/.hg/store/revlogmeta

This is done by something like:

  cd repo
  hg.old export : > /tmp/a.txt
  cd ..
  hg.new init repo2
  cd repo2
  hg.new import /tmp/a.txt
  # copy new data from repo2 to repo

The new test data is compatible with both the old and new versions of hg.

Reviewed By: muirdm

Differential Revision: D48019740

fbshipit-source-id: e6f97e4c236d1a0a35e290b864c462818923d603