Add onDismiss to Tooltip
Summary:
We'd like to be able to log when a tooltip is dismissed. Especially for component tooltips. It's not recommended to do this via "unmount" effects. So let's add a callback to the tooltip.

Edit: we can accomplish this using an effect that's looking at visibility, instead of calling it every time `setVisible(false)` is used

Reviewed By: quark-zju

Differential Revision: D48011236

fbshipit-source-id: ccf91dfe4c8afbcde88fcb4ef515c6c1c902de2b