status: kill EdenFS fastpath
Summary:
This is no longer in use and has been superseded by the "full" Rust status.

The fastpath status was relying on the dirstate parsing code in here to read out p1 as an argument to "getScmStatusV2", so I piped p1 through from a different direction instead. This means all "sl status" calls on EdenFS will parse the dirstate one fewer time.

Reviewed By: quark-zju

Differential Revision: D48090386

fbshipit-source-id: 54d97381af20ad3a3e5ab095f96251db27f338b7