Add heartbeat message and get log file
Summary:
Add a heartbeat message we can use to diagnose if the server is still reachable. We can use this to check for reachability in the bug menu to provide more resources.

We also want to be able to send the log file to the UI so we can tell the user where they should look for logs in the event of a crash (to self-diagnose)

Reviewed By: quark-zju

Differential Revision: D48167360

fbshipit-source-id: 5bb7141d5dc64dfa9b6578f7fb753501b29665fe