Add warning to help menu if server unreachable
Summary:
If you open the help menu, we should send a heartbeat and if it fails, tell the user where they can find their log file.

We do this with a simple `useHeartbeat` hook which sends a heartbeat and waits for the answer.

If the heartbeat comes back, then we show the help menu as normal

Otherwise, we render an error notice since it's a severe state to be in.

Technically, I expect that the main app connection status is sufficient for this heartbeat. But this is an extra check on top to prevent someone from doing something that ends up being useless like clicking the file a bug button.

Reviewed By: quark-zju

Differential Revision: D48167527

fbshipit-source-id: c0b416ded0937d443c782642037bdde1873d789c