Fix tab order of text area toolbar
Summary: If you have the title and summary both being edited, tabbing from the title should go to the summary next, not the attach file paperclip. This is an easy fix by changing the render order. The toolbar is positioned absolute, so its order in the DOM doesn't otherwise matter.

Reviewed By: sggutier

Differential Revision: D48199507

fbshipit-source-id: 8aa1efc0839049ec7c5024abc48e9d7db749af5a