Filter out duplicates when reporting NFS crawling
Summary:
When the NFS crawl detection logic runs, it walks over the list of processes that have open files in any of the Eden mounts. If a given process hierarcy has mulitple non-exluded process with open files they are reported multiple times. This change keeps track of which pids have been seen in the reporting and only reports them one time. It also detects process hierarchies that are proper subsets of and only retains the longest hiearchy for reporting.

NFS crawl detection is still pretty basic - simple detecting that certain metrics have crossed a threshold and reporting which processes have open files in Eden mounts. Future improvements could use this trigger to initiate sampling of open files over a short interval to better refine which processes are actually opening multiple files and not just holding a static list. This improvement could be made in the future if this simple mechansim is too noisy.

Reviewed By: mshroyer

Differential Revision: D48095393

fbshipit-source-id: c85f5dd7d8a7d184cd41d4b0604795874e123641