Implement Bundle generation
Summary: With `Packfile` generation in place, we can create a [Git bundle](https://git-scm.com/docs/bundle-format) as a self contained package that can be directly used by a client to get latest changes. The current test is very basic and just checks if the bundle gets generated as desired. Subsequent diffs will validate if the bundling workflow works as expected in an actual Git repository.

Differential Revision: D48010557

fbshipit-source-id: e6cd5cfa4aa043485753f3142a28ccacf7796c0f