Simplify NFS crawler process hierarchy logging
Summary:
Gather info about NFS crawlers using an intermediary tree, represented
abstractly as a map of PIDs to process information including their parent PIDs.

This simplifies things a little by making the properties of the output that we
care about--that it's a collection of the longest paths from leaf PIDs
accessing EdenFS mounts to their parents--follow from the structure of the
intermediate result we're bulding up rather than something we enforce
procedurally.

This way we also avoid duplicating calls to readProcessSimpleName and
proc_pidinfo via getProcessHierarchy.

Reviewed By: jdelliot

Differential Revision: D48219960

fbshipit-source-id: 3f0fb5c64ef16740252af0ff1234a20c08f2517f