Use succession tracking for previews
Summary:
Now that we track successions, we can use it in operation previews to make previews more correct.

To start with, let's add this for rebase and goto.

Imagine the following situation:
- you rebase a stack of commits A,B,C
- while it's running, you click "goto B"

Before, we would show "you're moving here" on commit B UNTIL the very first list of commits comes back after the rebase. During the time after the rebase but before the checkout is fully finished, we wouldn't render "you're moving here", because our test for where to write that label was looking for the OLD commit hash for commit B, but it was succeeded during the rebase.

To support this, I add recoil state of successions. Generally, most state doesn't want to use this map, but rather listen for successions to do a one-time migration of stored hashes. Previews don't easily fit this pattern, so instead we lookup successors while running.

Due to queued operations and previews being applied in order and successions only possibly applying up to the currently running operation, I don't think there's a situation where using successions will break preview appliers.

I add a util here that makes it easy to lookup succesors from the preview context. This way the preview appliers don't need to change much, and moreso reflect how we pass "SucceedableRevset" as arguments in the first place.

Basically, from now on, operations' preview appliers should basically never deal with hashes directly, but rather use `latestSuccessor()`.

Reviewed By: quark-zju

Differential Revision: D48133149

fbshipit-source-id: 0ffcd00e4006f08a912d281879cad59aae50cd51