global access to recoil get/set
Summary:
One problem I've been running into with Recoil is that it's hard to set state outside of the UI. If we get events from the server or some other async thing resolves at some point, it's hard to have that update multiple pieces of state.

The way recoil is designed, it encourages atom effects to only set themselves. This goes so far as even an atomFamily can't set other keys in the same family. Sure, you could have it be derived from another atom which is a Map, but then you're doubling up your Map for no reason.

The foremost use case is that we want succession events, which are triggered by the server returning new commit data, needs to cascade and affect many other atoms.
The flow is like this:
- the server sends a new list of commits
- the list of changes is parsed and sets the latest list of commits in the UI
- while parsing this list, we notice commits which have been succeeded
- we now want to do a ONE-TIME update of all other atoms which depend on this state. Because this is one-time, we can't model it as an atom. Even if we did, that atom couldn't affect other atoms.

I don't like that recoil seems incapable of handling this case. I haven't seen a compelling organizational reason that this is discouraged. Perhaps you could argue that having atoms effect each other can lead to messy design. But not every problem can be modeled with a perfect hierarchy of dependency. If I remodeled our application in this way, it would be far more complicated than if we just allowed some atoms to effect others, IMO. We'd need some weird global object that goes outside of recoil and then have atoms descend from there, and basically remove the whole point of recoil. Or have a ton of effects in render functions, which makes the state highly dependent on how it's rendered and hard to read.

I found one interesting library in OSS, [recoil-nexus](ttps://github.com/luisanton-io/recoil-nexus) that adds a backdoor to let you get/set any recoil atom/selector at any moment, using recoil primitives. This is basically as if we allowed atom effects to use `set(anyAtomName, value)`. We could use this to listen to arbitrary subscriptions, or have one atom's `onSet` call `set()` on other atoms to have atoms effect each other.

This diff brings a form of this library internally. I modified it slightly, but included the original copyright license.

Note that one reason that this hasn't been brought into upstream recoil is that this requires a sort of "global" handle on the recoil state. If you had multiple recoil roots, this could get kind of weird/invalid. We don't really have that situation in our codebase so I don't mind ignoring that case.

Reviewed By: quark-zju

Differential Revision: D48086215

fbshipit-source-id: 028a6b81179af104590da48269821cee7d6d5097