Fix download commit box input sources
Summary:
Someone reported general weirdness with the commit download input box, sometimes running with no arguments.

I think what was happening was related to using `onKeyUp` for input collection. If you paste with cmd-v, keep holding cmd, then click the button or press enter, the keyUp wouldn't have fired yet (but key down would have).

I think it's better to use `onInput`, so we are for sure syncing the input state of the text field. Then we can use keydown for the Enter press, to avoid weirdness with holding a modifier.

Reviewed By: quark-zju

Differential Revision: D48239163

fbshipit-source-id: 7fcfe959936607d3c0a9e3dd81391d83f2c29f13