Fail when importing non-UTF8 encoded refs
Summary:
The current implementation of `gitimport` imports `refs` by converting the ref names into `String`. However, Git allows for ref names to be non-UTF8 encoded and hence this conversion to `String` would not always be successful. To work around this, we lossily convert the ref name into a `String` replacing non-readable characters with `�` before importing the ref as a Mononoke bookmark. This approach can lead to a lot of problems (breaking the branch, confusing two different branches as one, etc.). For now we are changing `gitimport` to fail when it comes across any such branch instead of performing a lossy transformation. Long term we might want to support non-UTF8 branch names.

Note that `remote-gitimport` doesn't need to be modified since it expects the ref name to provided as a `String` input so it will automatically fail when provided with a non-UTF8 encoded byte sequence.

Reviewed By: mitrandir77

Differential Revision: D48234429

fbshipit-source-id: 96ea7ee91666cb4ece0b28f01a45c5cfb7ee80a3