Fixes to getdeps OSS builds on CentOS Stream 9
Summary:
CentOS Stream 9 comes with a much newer version of gcc (11.4.1) than either earlier devservers (8.5) and sandcastle (8.4).

This set of changes enables builds to succeed on newer devservers. One set of errors was realted to ImmediateFuture where `makeNotReadyImmediateFuture` failed with an "class template argument deduction failed" error. Looking at the errors it looked like using parens instead of curly braces might be the solution. However, this did not change the outcome. Another error in the output was that the `SemiFutureReadiness` enum was private. Moving this to public fixed this error.

Next up was an issue where size_t was undefined in RingBuffer.h. Easy add of <stddef.h> resolved this issue.

Reviewed By: xavierd

Differential Revision: D48250671

fbshipit-source-id: 646e0d5865593a16ba29131c2443b87f2bb4820e