Integration test for git-bundle
Summary: Now that we have the newadmin `git-bundle` command for creating bundles with our implementation, this diff adds an integration test that E2E validates that our bundling logic works as expected. With a normal Git repo, we create a Git bundle in code and then let the Git client verify and unpack the bundle. This test will catch regressions in our bundling logic if they get introduced in the future.

Differential Revision: D48163134

fbshipit-source-id: d41af644137a3982b4d750dbd0bd012d43bb8be0