Command for Git bundle creation
Summary: This diff creates a `git-bundle` command that takes as input a Git repository path and walks the Git directory, finds all the loose objects with refs and packs them into a bundle using the `BundleWriter` implementation. This command will be used in integration test to E2E validate that our recent Git changes work as expected.

Differential Revision: D48152494

fbshipit-source-id: 8ebac77d5f68a42405c25e683eb8ab9e4e44a9f7