tunables: make update container type a parameter
Summary:
We want to add convenience methods to allow us to treat tunables as unsigned.  This means we will want to make the container type different from the external type.  To make things simpler, move these types into the parameters for the generator function.

The by repo value type is itself unnecessary, as the compiler can infer this in the generated code.

Reviewed By: RajivTS

Differential Revision: D48265823

fbshipit-source-id: d6e33161115bb2a397a1ba171b390737a721c699