tunables: remove utility function and improve docs
Summary:
The `get_duration_from_tunables` utility function doesn't add anything: it makes the callsite longer, and because of the necessary customizations, seems harder to read than simple rust standard library methods.  Inline its calls to remove it.

Where it was used in the derived data manager, improve the documentation to make it clearer what these tunables are for.

Differential Revision: D48171717

fbshipit-source-id: e8084e474f2ac994a0905481390881595a218980