tunables: fix comments and error messages about tunable types
Summary:
We support more than just atomic bools and ints, but we require users to specify the type using the tunable type aliases.  Make this clear as a requirement.

The old error messages weren't updated because they were in a constant at the other end of the file.  For these messages, it's better to just have the message where the error occurs.

Reviewed By: mitrandir77

Differential Revision: D48265822

fbshipit-source-id: c0d2adf91247505a6e94051ec699eec336fb7c92