tunables: implement unsigned tunables
Summary:
We often want tunables to be unsigned, e.g. so that they can be passed to the `Duration` constructors.  In the thrift format they are `i64`, so we usually end up casting them at their usage (`value as u64`).

Make this more explicit and ergonomic by defining `TunableU64` and `TunableU64ByRepo` types, which perform the cast for us.  Note that negative values get cast to large positive integers, which means we can still represent very large u64 values if we want to, we'll just need to encode them when we generate the config.

Reviewed By: RajivTS

Differential Revision: D48265824

fbshipit-source-id: 9e182ec2ffe968edaa9949db3498d5d83b1bd344