tests: rewrite test-server-large-path-and-content
Summary:
This test originally tested that both mononoke and blobimport could correctly import large files and files with long paths or names.  It was particularly interested in edge cases in the revlog format to do with when revlogs were split into separate data and index files, and how very long filenames were encoded in the store.

The Mercurial client we use no longer uses the revlog format internally, and for the purposes of blobimport, we use `debugexportrevlog` which exports a much simpler revlog format for which these edge cases no longer exist.  As such, the test isn't really testing anything anymore.

It's still worthwhile for us to check that Mononoke can correctly receive and serve files or paths that are long, so we adapt this test to just test that.

Reviewed By: quark-zju

Differential Revision: D48352045

fbshipit-source-id: f1f045ca84090a4477972fd679c34c7f25123cd4