Constrain selection in comparison view to one column
Summary:
If you start dragging in the comparsion view, it would select across all columns (line number, before, line number, after), since it's implemented as a table. This is annoying if you want to copy/paste from one side, e.g. some lines in before.

The table is nice because it handles line wrapping smartly, allowing either left/right to wrap, and handles sizing properly. `display: grid` may also be workable, but I had performance and other issues when I tried that in the comparison view in the past. Other places seem to be using tables for this kind of rendering too.

To fix the selection issue, we need two things:
1. when you copy, find which column you had started your selection in, and use only the text content from that column
2. don't render the selection on any columns except the one you start clicking.

Reviewed By: quark-zju

Differential Revision: D48339812

fbshipit-source-id: 35f1a3dd5d82baab7652658b16455260a4b36804