util::path: add fn analagous to pathutil.canonpath
Summary:
I'm working towards supporting the path matcher in pure Rust commands. A key part of massaging CWD-relative patterns (e.g. "glob") in Python is pathutil.canonpath, which makes a path relative to the repo root given the repo root and CWD. In this commit I implement a similar method root_relative_path in Rust to serve a similar purpose.

It doesn't do any path auditing because we think the path auditing we do when writing files in the Rust VFS is sufficient.

Reviewed By: zzl0

Differential Revision: D48090375

fbshipit-source-id: 1d1182a4d0d4513c17278e2e5ab7545f23135a2a