move textmate-lib to shared/
Summary:
We're going to use textmate-lib for syntax highlighting in ISL, so let's move it to shared.

textmate-lib is the stuff that lets us actually do the syntax highlighting based on a set of grammars. See the textmate-lib readme file for more info.

Those grammars are fetched and setup via addons/textmate, which manages fetching actual grammars from github. To get syntax highlighting in ISL, we'll also need to do this download/codegen step for ISL, even though reviewstack already does it.

Reviewed By: quark-zju

Differential Revision: D48367396

fbshipit-source-id: e104d17b9385e1550d002487ac5b2f969aab3b3c