dirsync: persist dirstate callback across dirstate refreshes
Summary:
When turning on the Rust status, test-fb-ext-dirsync-absorb.t was failing because the Rust status was eagerly flushing the treestate, and dirsync's parent-change-callback was getting lost. Fix by wrapping localrepository.dirstate to always add the callback when a new dirstate is initialized.

I copied what the journal extension does to make the callback stick.

Reviewed By: zzl0

Differential Revision: D48340340

fbshipit-source-id: 3c4305be0d448d9e2dc1f6a8f17d338663096aa7