status: update treestate mtimes in non-watchman case
Summary:
Similar to the Watchman case, we now update treestate mtimes as a side effect of computing the working copy's pending changes. This is an optimization which avoids content comparison in future status calls.

We update the mtime when we get a ResolvedFileChangeResult::No for a file. This means we compared the content, and things match. Since we resorted to compared the content, we can be pretty sure the treestate's previous mtime didn't match (or wasn't set at all).

Reviewed By: quark-zju

Differential Revision: D48117905

fbshipit-source-id: 5a3d2a11a592382d44ac67945a5ed033207a3bbb