clone: stop using Url for "source"
Summary:
I'm going to make the Rust clone support file path sources, and the Url type makes things annoying. We don't really want a Url anywhere, anyway. paths.default is a string, so everything already prefers that.

In this commit I add a struct to hold the bits of info about the clone source. Shouldn't be any functional changes.

Reviewed By: quark-zju

Differential Revision: D48117892

fbshipit-source-id: 6a7e4518236ebd14b4573e97ad1290f2dfbcd4c2