tools/admin: move repo-info back to its own command
Summary:
Unwind the refactor in D41401749 and move `repo-info` back to a top-level command.

Commands in the admin tool need to be discoverable, which is harder to do when things are hidden under duplicative top-level names like `repo` and `repos`.  Instead, the top level commands should be fairly clear as to what they do, and either refer to an action (like `repo-info`) or an obvious category of things (like `locking`).

Reviewed By: mitrandir77

Differential Revision: D48191619

fbshipit-source-id: dc689d9fb8471c6a68fa156a39d45df4d171da61