Respect case for enum in MySQL
Summary:
Before this change, the "stage" would be inserted into prod SQL as an empty string as the case of the enum didn't match between the enum and the db schema.
This would lead to future queries that tried to select a specific stage to find no matching entries.

For that reason, we could `commit` the plan to the db, but we wouldn't then be able to `execute` the plan.

Reviewed By: RajivTS

Differential Revision: D48357385

fbshipit-source-id: 72532e1c6ff7413cac0adcbd57e79afc67c5787a