Import HEAD symref as part of Gitimport
Summary:
So far `gitimport` only imports the direct refs, i.e. `branches` and `tags`. However, when generating bundles and packs, we also need the value of the `HEAD` symref. This diff adds support to import the `HEAD` symref based on CLI args. Currently, the integration tests for `gitimport` are not passing this flag. Once I have a `newadmin` command to get/set symrefs, I will update the integration tests to ensure things work as expected.

Note that reliably importing `HEAD` symref via `remote-gitimport` is not possible since the hook gets triggered conditionally and there are ways that the hook can be bypassed. A better approach would be to allow `HEAD` changes to only go through Mononoke admin tools which would ensure that the change is always tracked in Mononoke as well as Git.

Differential Revision: D48349068

fbshipit-source-id: 67bc08625cb4a65ebe2164c6afd8bab2c35a710b