Introduce git_symbolic_refs for tracking symrefs
Summary:
In Git, in addition to regular refs under `refs/` folder, a user can also create a `symref`. In contrast to regular refs, `symrefs` point to another ref instead of directly pointing to a commit. So far, both `gitimport` and `remote-gitimport` ignore `symrefs` but while creating bundles we will need to atleast include `HEAD` which is a `symref` that gets created by default in every git repo.

Once this diff is landed, I will update `gitimport` to store the `HEAD` `symref` as part of the repo import.

Differential Revision: D48315266

fbshipit-source-id: 1ff47be09488197b2ef93d9844a02a37c4a396f1