Implement factory methods for git_symbolic_refs
Summary: The previous diff created a table for `git_symbolic_refs` and also introduced it as a new `facet`. This diff implements the factory constructor method for this `facet` so that it can be used directly in existing facet containers. The next diff will make use of this `facet` in `gitimport`

Differential Revision: D48316244

fbshipit-source-id: b27661e0178599152ed0fd5600a8efa164cc01f9