clone: support "--updaterev" in Rust
Summary:
We now use the basic Rust revset resolution to resolve the "-u" checkout destination. The only things that will work are bookmarks in selectivepulldefault, or commit hashes that are ancestors of those bookmarks.

If we end up needing full revset evaluation and/or autopull, we will need to fall back to Python.

This behavior is behind a config flag so I can roll it out slowly.

Reviewed By: quark-zju

Differential Revision: D48117906

fbshipit-source-id: 5e601e0bc27c59d1edd454eda1e77b43f6a01f6c