debugruntest: hacky support for #testcases
Summary:
Implement poor man's #testcases from run-tests.py.

"#if <testcase>" is replaced with a runtime check instead of a "compile time" check as with other features.

The runner iterates over test cases and sets a variable which the transformed "#if <testcase>" checks to see if it is in that case.

The output/diff has no knowledge of the test cases, so it's annoying to debug failures. I'm sure we can improve that.

I ran into some issues with pydoceval. It was previously relying upon being run at the module level where locals() and globals() are the same thing. I bungled my way around it.

test-fb-ext-smartlog-smartdate.t had accidentally been migrated to debugruntest with testcases. It uses "exit 80" to bail out of an unsupported testcase early, which debugruntest doesn't support, so un-debugruntest it.

Reviewed By: quark-zju

Differential Revision: D48073293

fbshipit-source-id: 4d13d7f0a266f2189b2e5c2eb0679db514661bcd