start of logging to scuba with out scribe libs
Summary:
eden cli in dev mode takes a really long time like ~15s for even the most simple
commands. A lot of this time is spent in linking. And out most expensive
dependency is the scribe logger. Let's do our own logging to cut this dependency.
It makes eden dev run in less than a second.

Reviewed By: xavierd

Differential Revision: D47884467

fbshipit-source-id: ba1ee3bc25e70f8e34ac8de333c59250e0cabdc6