Add git-symref command in Mononoke newadmin
Summary: `gitimport` now captures symbolic refs while importing a repo. This diff adds a `newadmin` command for fetching, adding, updating and deleting a symbolic ref in Mononoke. This can be used when updating the HEAD symrefs on a Git repo to keep its Mononoke counterpart in sync. For now, it will be used in integration tests.

Differential Revision: D48357780

fbshipit-source-id: df09900d36453901bf3cf7b9c41709594ea232d6