pathmatcher: add HintedMatcher to support Python hints
Summary:
Python has fast paths that depend on various hints from the matcher. For example, if the matcher always matches, it skips matching altogether.

HintedMatcher decoretes a Rust matcher with the requisite hints. It only makes sense to use at the top level of matcher creation (e.g. the positional CLI pattens, the --include patterns, and the --exclude patterns).

Reviewed By: zzl0

Differential Revision: D48090395

fbshipit-source-id: 70147a2db5db179592f2d2ec9d0469b7453315a4