clean: support using Rust status with "--dirs"
Summary:
"--dirs" needs to search for all empty directories. Previously for non-Watchman it would set a callback on the matcher to record all directories during the walk. That didn't work with the Rust status since random callbacks on the matcher are not supported in the Python->Rust matcher conversion.

Fix by changing "--dirs" to just do a separate walk to find empty directories.

Note that this also fixes the "--dirs" option for EdenFS (but at a terrible cost). I'll follow up with a check that requires a pattern to avoid full repo walk.

Reviewed By: quark-zju

Differential Revision: D48117913

fbshipit-source-id: 43948e1f8d8043f6002aa63847f6c73513c2d68d