workingcopy: get rid of PendingChangeResult layer
Summary: This only existed so that status could propagate directories it found during the fs walk, but this doesn't make sense to be part of the core "PendingChanges" trait. I think the only potential user of this was "sl clean --dirs", but I already changed that to use the Rust walker directly instead of rely on "status" to do the fs walk.

Reviewed By: quark-zju

Differential Revision: D48117896

fbshipit-source-id: 2b53b7c68fcbafa7e3b3ceba1177899d8d46580b