status: support "--ignored" in Rust for watchman
Summary: Support outputting ignored files for the Rust+watchman case. If ignored file tracking is enabled, we do a normal watchman query (i.e. not a "fresh" query), and spit out any paths we see that match the ignore matcher. If ignored file tracking is disabled, we trigger a fresh watchman query since ignored files will not be present in the treestate.

Reviewed By: quark-zju

Differential Revision: D48340336

fbshipit-source-id: 0d50542a614d9e6dcb43c99cc402c39d852fbed8