status: support "--ignored" for non-watchman case
Summary:
Respect the "include ignored files" option in the Rust "physical" filesystem (i.e. non-EdenFS, non-watchman). When walking files we first check if the file matches the ignore matcher, and if so propagate as an ignored file.

This works because the ignore matcher does not match files in the manifest (i.e. it won't report a checked-in ignored file as ignored).

Reviewed By: zzl0

Differential Revision: D48340335

fbshipit-source-id: 5b4b2fe59540937ed58cea9d8b9c75f83216819e