status: enable Rust status in test-rebase-interruptions.t
Summary:
This test was deadlocking because it had a precommit hook that ran "hg status" and was waiting for the working copy lock to flush changes the dirstate.

To mitigate this issue, the Rust dirstate flushing now approximates the Python logic of "only wait for working copy lock if we have lots of changes to write". Rust achieves this by simply counting the number of pending updates to the treestate. I picked some arbitrary thresholds. The important thing is if there are a lot of updates, you wait for the lock since not updating will cause performance degredation (since subsequent "status" invocations will redo the work). If there aren't a lot of changes, don't wait (and don't update the dirstate) since the updates "status" does are just to optimize future performance.

Note that the deadlock is still possible, just less likely.

Reviewed By: quark-zju

Differential Revision: D48117914

fbshipit-source-id: b36c5327a14ae208c75876d77e9232e684b5ec6c