Made highlighted commits unhighlight properly
Summary:
We highlight commits when hovering a button that affects them, such as "resubmit stack". This wasn't working well if you click the button, since the highlights seem to get stuck.

I was using onMouseLeave, but it seems onMouseOut is more reliable for this. seems like if the rendered child changes, then the mouse leave is not properly updated, but it works well with onMouseOut.

Reviewed By: muirdm

Differential Revision: D48415221

fbshipit-source-id: 564268cceb2fe595f6fad866dc823880d79d756a