Add minimum drag target width so rebase drag & drop is easy on small commits
Summary:
When dragging a commit to rebase, it was possible to "miss" a target commit entirely. This is because the drag target was only set on the commit itself, so only as wide as the commit is. We can't just make commits span the whole width, because that causes weird behavior with buttons and such.

Instead, we can do a sort of a hack where we add a second absolutely positioned div that acts as an explicit drag target. We can choose any arbitrary width to make this drag target.

Initially I wanted it to be 100% of the width, but this actually doesn't work well with absolutely positioned elements, since that enables scrolling horizontally.

instead, we can just set it to 500px arbitrarily. This does mean that the commit tree list now essentially has a min-width of 500px, but I'm not so opposed to that honestly. That's kind of already the case thanks to commit title + buttons like goto.

Reviewed By: muirdm

Differential Revision: D48413491

fbshipit-source-id: f31139b904dff89b5785e7c855ee6cdea77342d0