pathmatcher: support already expanded filesets
Summary:
In order to support filesets in the Rust matcher, Python will expand the file sets to lists of files, and then pass them to Rust to incorporate into the Rust matcher.

This commit adds support to the HintedMatcher for the already-expanded filesets. Basically, they are just mixed in as an exact matcher.

Reviewed By: zzl0

Differential Revision: D48090387

fbshipit-source-id: d2a0693e8d959e617f12a4af5669a6b49cd0bbf9