Start with some files collapsed by default in comparison view
Summary:
If you have a diff with dozens of files, it can be slow to load in the comparison view. If we start excess files as collapsed by default, it will load a little faster. You can still expand files manually of course. Of course we still have to wait for `sl diff`, but the rendering and syntax highlighting tax will be much lower.

We can also be a little tricky and collapse files with lots of changes. For example, if you added a 7000 line file, we can collapse that by default to avoid rendering and syntax highlighting. Most people probably don't actually want to scroll through that whole file. But if they do, all they have to do is expand the file.

We'll probably also want an "Expand all" button, but I'll do that separately.

Reviewed By: muirdm

Differential Revision: D48404044

fbshipit-source-id: 4626dc0cbe7824c0d024b33099bfbc8c0b019a37