Add syntax highlighting for loaded context lines
Summary:
You can load extra missing lines of context in the comparison view. These are not part of the ParsedDiff, so we need to highlight it separately. To do this, I added a second API for syntax highlighting a chunk of file content.

I also noticed there's a tokenizeLines api in textmate-lib, so we can skip a `.join('\n')` on the file contents, which should reduce memory usage especially for large files.

This method is not too bad for performance, even for large amounts of context, because we don't auto-load context. It's a manual procedure, and it already takes async time to load from the server. So waiting a little longer for highlighting is generally not noticeable.

Plus since it's not a diff view, it highlights half as much as an equivalently sized diff hunk.

Reviewed By: muirdm

Differential Revision: D48402052

fbshipit-source-id: e3b5564c974aa0cb6b6577bbb90d530e2c719365