Introduce cancellation token
Summary:
Add a cancellation token class to `shared/`.

This is useful for cancelling async work. VS Code uses the same concept and it's pretty good at this job.

Use case at hand is that react effects render twice, to check for mount/unmount bugs in your effects. Effects should be able to clean themselves up. A cancellation token makes this easy to implement:

```
useEffect(() => {
  const token = new CancellationToken();
  doMyAsyncWork(token);

  return () => token.cancel();
});
```

(used this way in the next diff)

Reviewed By: muirdm

Differential Revision: D48400651

fbshipit-source-id: 699b3db4c4bf3714225e877367eb18916f6afe6d