Performance improvements for syntax highlighting
Summary:
I dug into the performance of the comparison view's syntax highlighting. It's not bad, but I found some ways to speed things up.

- if we hoist ensureOniguramaIsLoaded, getGrammar store doesn't need to be async. This fixes cachedGrammarStore not properly caching when multiple requests fire at once.
- we can add a cache to getGrammar, keyed by scopeName. This way all files of the same type (e.g. typescript) don't redundantly call loadGrammar. So you only pay once per file type. Note that this requires invalidating the cache when the store changes, since the colors change.
- Add cancellationToken to `useEffect` that actually calls tokenizeHunks. This effect is most pronounced in dev mode, where the effect runs twice on startup, to detect effect cancellation bugs. The cancellation token allows us to bail on the syntax highlighting work if the component unmounts quickly.

Originally I intended to move highlighting to a web worker. IT seems the current approach with vscode-onigurama is actually pretty fast, at least for common small-ish diffs. I think it helps that we only highlight hunks, and not the whole file. Large diffs may still be affected, so we should do more measurements on larger diffs and consider still moving to a web worker.

The comparison view does cache the last change it was looking at, which means if you close and reopen the view, it will highlight once, then refetch the diff from the server and highlight it again since it will be a new value. I don't think this is a very big deal, but on very large diffs it may be a noticeable extra perf hit. I have a couple of ideas to improve this:
- only try to syntax highlight files that are in the view, so you have to scroll down to syntax highlight lots of files
- if there's LOTS of files, start most of them collapsed so they don't highlight at all.

Reviewed By: muirdm

Differential Revision: D48400650

fbshipit-source-id: 69ba7c0dbe56786208f184f51e01a58d76f32212